refactor: move data source specific functions to ConnectionPool class
connection pool will take care to close sessionfactory and datasource
