+++ src/main/java/io/supertokens/storage/sql/hibernate/HibernateUtils.java
@@ -31,6 +31,7 @@ import org.hibernate.cfg.Environment;
 import org.jetbrains.annotations.NotNull;
 
 import javax.sql.DataSource;
+import javax.xml.crypto.Data;
 import java.io.Closeable;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -39,11 +40,9 @@ import java.util.Properties;
 
 public class HibernateUtils {
 
-    private static final List<Closeable> CLOSEABLES = new ArrayList<>();
-
-    public static SessionFactory initSessionFactory(DatabaseConfig databaseConfig) {
+    public static SessionFactory initSessionFactory(DatabaseConfig databaseConfig, DataSource dataSource) {
         final StandardServiceRegistry serviceRegistry = new StandardServiceRegistryBuilder()
-                .applySettings(properties(databaseConfig)).build();
+                .applySettings(properties(databaseConfig, dataSource)).build();
 
         final MetadataSources metadataSources = new MetadataSources(serviceRegistry);
 
@@ -58,8 +57,6 @@ public class HibernateUtils {
 
         SessionFactory sessionFactory = sessionFactoryBuilder.build();
 
-        CLOSEABLES.add(sessionFactory);
-
         return sessionFactory;
     }
 
@@ -72,7 +69,7 @@ public class HibernateUtils {
         return new CustomNamingStrategy(databaseConfig);
     }
 
-    private static Properties properties(DatabaseConfig databaseConfig) {
+    private static Properties properties(DatabaseConfig databaseConfig, DataSource dataSource) {
         Properties properties = new Properties();
         // TODO: sql-plugin: take from config or let hibernate figure out what driver is needed
         // TODO: sql-plugin -> even if I give MySQLDialect when using postgres, tests still pass. Is this done
@@ -87,86 +84,9 @@ public class HibernateUtils {
         }
 
         // data source settings
-        DataSource dataSource = dataSource(databaseConfig);
         properties.put(Environment.DATASOURCE, dataSource);
 
         return properties;
     }
 
-    private static DataSource dataSource(DatabaseConfig databaseConfig) {
-        HikariDataSource hikariDataSource = new HikariDataSource(hikariConfig(databaseConfig));
-
-        CLOSEABLES.add(hikariDataSource);
-
-        return hikariDataSource;
-    }
-
-    @NotNull
-    private static HikariConfig hikariConfig(DatabaseConfig databaseConfig) {
-
-        HikariConfig hikariConfig = new HikariConfig();
-        hikariConfig.setJdbcUrl(getJdbcUrl(databaseConfig));
-
-        // TODO: sql-plugin -> choose the right driver based on actual config
-        // TODO: sql-plugin: do we need a validate function to check if the dialect passed exists or not
-        // like invoke a Class.forName() and if it throws an exception we do a quit program exception,
-        // or do we let hibernate handle if a dialect does not exist ?
-        // this basically might be a measure against spelling mistakes or such
-        hikariConfig.setDriverClassName(databaseConfig.getDriverClassName());
-
-        if (databaseConfig.getUser() != null) {
-            hikariConfig.setUsername(databaseConfig.getUser());
-        }
-
-        if (databaseConfig.getPassword() != null && !databaseConfig.getPassword().equals("")) {
-            hikariConfig.setPassword(databaseConfig.getPassword());
-        }
-        hikariConfig.setMaximumPoolSize(databaseConfig.getConnectionPoolSize());
-        hikariConfig.setConnectionTimeout(5000);
-        hikariConfig.addDataSourceProperty("cachePrepStmts", "true");
-        hikariConfig.addDataSourceProperty("prepStmtCacheSize", "250");
-        hikariConfig.addDataSourceProperty("prepStmtCacheSqlLimit", "2048");
-        // TODO: set maxLifetimeValue to lesser than 10 mins so that the following error doesnt happen:
-        // io.supertokens.storage.postgresql.HikariLoggingAppender.doAppend(HikariLoggingAppender.java:117) |
-        // SuperTokens
-        // - Failed to validate connection org.mariadb.jdbc.MariaDbConnection@79af83ae (Connection.setNetworkTimeout
-        // cannot be called on a closed connection). Possibly consider using a shorter maxLifetime value.
-        hikariConfig.setPoolName("SuperTokens");
-
-        return hikariConfig;
-    }
-
-    @NotNull
-    private static String getJdbcUrl(DatabaseConfig userConfig) {
-        String scheme = userConfig.getConnectionScheme();
-
-        String hostName = userConfig.getHostName();
-
-        String port = userConfig.getPort() + "";
-        if (!port.equals("-1")) {
-            port = ":" + port;
-        } else {
-            port = "";
-        }
-
-        String databaseName = userConfig.getDatabaseName();
-
-        String attributes = userConfig.getConnectionAttributes();
-        if (!attributes.equals("")) {
-            attributes = "?" + attributes;
-        }
-
-        return "jdbc:" + scheme + "://" + hostName + port + "/" + databaseName + attributes;
-    }
-
-    public static void close() {
-        for (Closeable CLOSEABLE : CLOSEABLES) {
-            try {
-                CLOSEABLE.close();
-            } catch (IOException ignored) {
-                // as of now the two closable we have (SessionFactory and HikariDataSource)
-                // do not throw IOException in their implementations
-            }
-        }
-    }
 }
